+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### 🛠️ Fixes
 - Fixed incorrect source branch name when creating a `chore` branch ([#202](https://github.com/candoumbe/pipelines/issues/202))
+- Fixed the behavior of `chore` command when running from a `chore/*` branch : the command will now properly end the `chore` workflow.
 
 ### 🚀 New features
 - Added `IDoChoreWorkflow` to `IGitFlow` and `IGitHubFlow`: this enable the `chore` target on those workflows